# MM-23547: remove redundant JSON parsing in push notifications