#### Summary
<!--
A description of what this pull request does.
-->


We remove the redundant JSON serialization and de-serialization
and use a DeepCopy method to accomplish that. The gains are significant.

```
name                                 old time/op    new time/op    delta
SendPushNotificationToAllSessions-8    6.55µs ± 1%    1.60µs ± 1%  -75.59%  (p=0.000 n=9+9)

name                                 old alloc/op   new alloc/op   delta
SendPushNotificationToAllSessions-8    3.53kB ± 0%    1.78kB ± 0%  -49.67%  (p=0.000 n=10+10)

name                                 old allocs/op  new allocs/op  delta
SendPushNotificationToAllSessions-8      23.0 ± 0%       7.0 ± 0%  -69.57%  (p=0.000 n=10+10)
```

We use a custom benchmark to check this. Because the mock store is so heavy
it takes up the entire profile. And the logging spams the console.

This is the benchmark I have used to measure it.

```
// Run it with | grep -v '{"level"' to prevent spamming the console.
// To run this, we need to minimize interaction of other components.
// Otherwise the benchmarks pollute too much.
// We hardcode to some sessions, to avoid calling mock store
// -       sessions, err := a.getMobileAppSessions(userId)
// -       if err != nil {
// -               return err
// +       sessions := []*model.Session{
// +               {
// +                       Id:        "id1",
// +                       UserId:    "user1",
// +                       DeviceId:  "test1",
// +                       ExpiresAt: model.GetMillis() + 100000,
// +               },
// +               {
// +                       Id:        "id2",
// +                       UserId:    "user1",
// +                       DeviceId:  "test2",
// +                       ExpiresAt: model.GetMillis() + 100000,
// +               },
// +      }
// And we set logger.zap = zap.NewNop() in mlog.NewLogger
// And we also comment out the call to a.sendToPushProxy(tmpMessage, session).
func BenchmarkSendPushNotificationToAllSessions(b *testing.B) {
	th := SetupWithStoreMock(b)
	defer th.TearDown()

	handler := &testPushNotificationHandler{
		t:        b,
		behavior: "simple",
	}
	pushServer := httptest.NewServer(
		http.HandlerFunc(handler.handleReq),
	)
	defer pushServer.Close()

	b.ResetTimer()
	msg := &model.PushNotification{
		Type:       model.PUSH_TYPE_MESSAGE,
		PostId:     "test",
		SenderId:   "userID",
		ChannelId:  "channelID",
		Category:   model.CATEGORY_CAN_REPLY,
		Version:    model.PUSH_MESSAGE_V2,
		IsIdLoaded: true,
	}
	for i := 0; i < b.N; i++ {
		err := th.App.sendPushNotificationToAllSessions(msg, "userID", "id1")
		if err != nil {
			b.Logf("fail: %v", err)
		}
	}
}
```

<!-- Thank you for contributing a pull request! Here are a few tips to help you:

1. If this is your first contribution, make sure you've read the Contribution Checklist https://developers.mattermost.com/contribute/getting-started/contribution-checklist/
2. Read our blog post about "Submitting Great PRs" https://developers.mattermost.com/blog/2019-01-24-submitting-great-prs
3. Take a look at other repository specific documentation at https://developers.mattermost.com/contribute
-->


#### Ticket Link
<!--
If this pull request addresses a Help Wanted ticket, please link the relevant GitHub issue, e.g.

  Fixes https://github.com/mattermost/mattermost-server/issues/XXXXX

Otherwise, link the JIRA ticket.
-->
https://mattermost.atlassian.net/browse/MM-23547