@agnivade 2020-04-07
I had marked the QA Testing Areas to be push notifications. Testing around that area should be sufficient.