@lindy65 2020-04-07
Ah, apologies - I missed that but just saw it on the ticket.  Is there any difference I should notice regarding push notifications or should they just work as per usual / expected?