@mattermod 2020-04-07
Creating a new SpinWick test server using Mattermost Cloud.