@mattermod 2020-04-07
Mattermost test server created! :tada:

Access here: https://mattermost-server-pr-14181.test.mattermost.cloud

| Account Type | Username | Password |
|---|---|---|
| Admin | sysadmin | Sys@dmin123 |
| User | user-1 | User-1@123 |