@mattermod 2020-04-08
Test server destroyed